differentiate between authentication and authorization routes
