+++ routes/index.js
@@ -1,10 +1,12 @@
 const { Router } = require("express");
 const SwaggerController = require("../controllers/SwaggerController");
-const authRoute = require("./authRoute");
+const authenticationRoute = require("./authenticationRoute");
+const authorizationRoute = require("./authorizationRoute");
 
 const router = Router();
 
 router.use("/docs", SwaggerController.serve, SwaggerController.setup);
-router.use("/auth", authRoute);
+router.use("/auth", authenticationRoute);
+router.use("/authorize", authorizationRoute);
 
 module.exports = router;
